requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
uvicorn==0.32.0
httpx==0.27.2
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
APScheduler==3.10.4
redis==5.0.8
fastapi==0.115.2
aiosqlite==0.20.0
pytest-asyncio==0.24.0
